fix: Fix heisenbug in gdb_sim.sh
Sometimes when running too fast in gdb it locked itself and returned
with error 131.
gdb_sim was reworked. I believe the cause was `set -o pipefail`
that caused bash not to track all possible childs in a pipe-chain
which caused bash to terminate one of the childs too early,
so the other child returned with SIGPIPE.
Anyway, should be working now.