refactor: Use GNUC __extension__ and remove argsizes from release
gcc is currently not properly handling folding and optimizing
when using compund literals. Merging same compund literals is not
working properly

So on __GNUC__ compiler i'll just use

    __extension({ static const variable ...; &variable })

and have put it all in a constant block, to minimize memory consumption

Moreover, this patch also removes argisizes from adding to printf
when compiling in release mode to reduce size (it's not really needed,
this is mostly a precoution).