feature(make): Fix checking floating sizes when compiling with -flto
Compiling with -flto causes the static library to be completely removed
which causes strings to not detect the part of the size inside the testing
when cmake tests the size of a type with check_type_size.
To change this behavior, I'll add -fno-lto flag for now to
CMAKE_REQUIRED_FLAGS so that the file for testing is compiled without
lto.